+++ timesketch/frontend-ng/src/components/Analyzer/TimelineSearch.vue
